requirements.txt
oracledb==2.0.0
mysql-connector-python==8.2.0

